revise model.py, model_result.js
