+++ model_test.py
@@ -1,3 +1,5 @@
+import sys
+import argparse
 import numpy as np
 import pandas as pd
 import matplotlib.pyplot as plt
@@ -5,12 +7,14 @@ import matplotlib.pyplot as plt
 # interpreter  python 3.7.2
 
 def pose_classification (list) : 
-    # print(list)
+    print(list)
     result = 'harry'
     return result
 
-def main (argv) :
-    result = pose_classification(argv)
+def main () :
+    #get parameter
+    ag = sys.argv[1:]
+    result = pose_classification(ag)
     print("result : "+ result)
     return result
